# Crud_FilmesAPI/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Crud_FilmesAPI",
    "version": "v1"
  },
  "paths": {
    "/api/filmes": {
      "get": {
        "tags": [
          "Filmes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Filme"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Filmes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Filmes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "format": "int32",
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/filmes/{id}": {
      "get": {
        "tags": [
          "Filmes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Filmes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Filme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Filmes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/generos": {
      "get": {
        "tags": [
          "Generos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Genero"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Generos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        }
      }
    },
    "/api/generos/{id}": {
      "get": {
        "tags": [
          "Generos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Generos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Genero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Generos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/locacaos": {
      "get": {
        "tags": [
          "Locacaos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Locacao"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Locacaos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Locacao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Locacao"
            }
          }
        }
      }
    },
    "/api/locacaos/{id}": {
      "get": {
        "tags": [
          "Locacaos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Locacao"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Locacaos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Filme": {
      "required": [
        "nome"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "dataCriacao": {
          "format": "date-time",
          "type": "string"
        },
        "ativo": {
          "type": "string"
        },
        "generoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Genero": {
      "required": [
        "nome"
      ],
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "dataCriacao": {
          "format": "date-time",
          "type": "string"
        },
        "ativo": {
          "type": "string"
        }
      }
    },
    "Locacao": {
      "required": [
        "cpf",
        "dataLocacao",
        "filmes"
      ],
      "type": "object",
      "properties": {
        "filmes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Filme"
          }
        },
        "cpf": {
          "type": "string"
        },
        "dataLocacao": {
          "format": "date-time",
          "type": "string"
        }
      }
    }
  }
}